Refactor for better organization.
