+++ src/gui/mod.rs
@@ -1,18 +1,16 @@
 //! Module for the GUI components of the application.
 
-use std::rc::Rc;
-
 use cairo::{Context, FontExtents, FontFace, FontSlant, FontWeight, Matrix, Operator};
 use gdk::{keyval_from_name, EventButton, EventConfigure, EventKey, EventMotion, EventScroll,
           ScrollDirection};
 use gtk::prelude::*;
-use gtk::{CheckMenuItem, DrawingArea, Menu, MenuItem, RadioMenuItem, SeparatorMenuItem};
+use gtk::{DrawingArea, Menu};
 
 use sounding_analysis::Layer;
 use sounding_analysis::layers::{warm_temperature_layer_aloft, warm_wet_bulb_layer_aloft};
 use sounding_base::DataRow;
 
-use app::config::{ParcelType, Rgba};
+use app::config::{Rgba};
 use app::{AppContext, AppContextPointer};
 use coords::{convert_pressure_to_y, DeviceCoords, DeviceRect, Rect, ScreenCoords, ScreenRect,
              XYCoords};
@@ -34,33 +32,8 @@ pub use self::text_area::update_text_highlight;
 
 use self::utility::{set_font_size, DrawingArgs};
 
-macro_rules! make_heading {
-    ($menu:ident, $label:expr) => {
-        let heading = MenuItem::new_with_label($label);
-        heading.set_sensitive(false);
-        $menu.append(&heading);
-    };
-}
-
-macro_rules! make_check_item {
-    ($menu:ident, $label:expr, $acp:ident, $check_val:ident) => {
-        let check_menu_item = CheckMenuItem::new_with_label($label);
-        check_menu_item.set_active($acp.config.borrow().$check_val);
-
-        let ac = Rc::clone($acp);
-        check_menu_item.connect_toggled(move |button| {
-            ac.config.borrow_mut().$check_val = button.get_active();
-            ac.mark_data_dirty();
-            ac.update_all_gui()
-        });
-
-        $menu.append(&check_menu_item);
-    };
-}
-
 pub fn initialize(app: &AppContextPointer) -> Result<(), SondeError> {
     sounding::SkewTContext::set_up_drawing_area(&app)?;
-    build_sounding_area_context_menu(&app)?;
     hodograph::HodoContext::set_up_drawing_area(&app)?;
     control_area::set_up_control_area(&app)?;
     text_area::set_up_text_area(&app)?;
@@ -94,85 +67,6 @@ pub fn show_pop_up_menu(app: &AppContext, _evt: &EventButton) {
         // menu.popup_at_pointer(ev);
         menu.popup_easy(3, 0)
     }
-}
-
-// FIXME: move these to the sounding module and add to set_up_drawing_area?
-fn build_sounding_area_context_menu(acp: &AppContextPointer) -> Result<(), SondeError> {
-    let menu: Menu = acp.fetch_widget("sounding_context_menu")?;
-
-    build_active_readout_section_of_context_menu(&menu, acp);
-    menu.append(&SeparatorMenuItem::new());
-    build_parcel_section_of_context_menu(&menu, acp);
-    menu.append(&SeparatorMenuItem::new());
-    build_profiles_section_of_context_menu(&menu, acp);
-
-    menu.show_all();
-
-    Ok(())
-}
-
-fn build_active_readout_section_of_context_menu(menu: &Menu, acp: &AppContextPointer) {
-    make_heading!(menu, "Active readout");
-    make_check_item!(menu, "Show active readout", acp, show_active_readout);
-    make_check_item!(menu, "Draw sample parcel", acp, show_sample_parcel_profile);
-}
-
-fn build_parcel_section_of_context_menu(menu: &Menu, acp: &AppContextPointer) {
-    use app::config::ParcelType::*;
-
-    make_heading!(menu, "Parcel Type");
-
-    let sfc = RadioMenuItem::new_with_label("Surface");
-    let mxd = RadioMenuItem::new_with_label_from_widget(&sfc, "Mixed Layer");
-    let mu = RadioMenuItem::new_with_label_from_widget(&sfc, "Most Unstable");
-
-    let p_type = acp.config.borrow().parcel_type;
-    match p_type {
-        Surface => sfc.set_active(true),
-        MixedLayer => mxd.set_active(true),
-        MostUnstable => mu.set_active(true),
-    }
-
-    fn handle_toggle(button: &RadioMenuItem, parcel_type: ParcelType, ac: &AppContextPointer) {
-        if button.get_active() {
-            ac.config.borrow_mut().parcel_type = parcel_type;
-            ac.mark_data_dirty();
-            ac.update_all_gui();
-        }
-    }
-
-    let ac = Rc::clone(acp);
-    sfc.connect_toggled(move |button| {
-        handle_toggle(button, Surface, &ac);
-    });
-
-    let ac = Rc::clone(acp);
-    mxd.connect_toggled(move |button| {
-        handle_toggle(button, MixedLayer, &ac);
-    });
-
-    let ac = Rc::clone(acp);
-    mu.connect_toggled(move |button| {
-        handle_toggle(button, MostUnstable, &ac);
-    });
-
-    menu.append(&sfc);
-    menu.append(&mxd);
-    menu.append(&mu);
-
-    menu.append(&SeparatorMenuItem::new());
-
-    make_heading!(menu, "Parcel Options");
-    make_check_item!(menu, "Show profile", acp, show_parcel_profile);
-    make_check_item!(menu, "Fill CAPE/CIN", acp, fill_parcel_areas);
-}
-
-fn build_profiles_section_of_context_menu(menu: &Menu, acp: &AppContextPointer) {
-    make_heading!(menu, "Profiles");
-    make_check_item!(menu, "Temperature", acp, show_temperature);
-    make_check_item!(menu, "Wet bulb", acp, show_wet_bulb);
-    make_check_item!(menu, "Dew point", acp, show_dew_point);
-    make_check_item!(menu, "Wind", acp, show_wind_profile);
 }
 
 trait Drawable: PlotContext + PlotContextExt {
